pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "blink-sync-sentry"
version = "0.1.0"
description = "Monitor Blink Sync Module local storage and automatically clean up old video clips."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "blink-sync-sentry contributors"},
]
keywords = ["blink", "camera", "local-storage", "cleanup"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
]
dependencies = [
    "blinkpy>=0.22.0",
    "aiohttp>=3.9",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
rich = ["rich>=13.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov",
    "ruff",
]

[project.scripts]
blink-sync-sentry = "blink_sync_sentry.cli:main"

[project.urls]
Homepage = "https://github.com/ballle98/blink-sync-sentry"
Repository = "https://github.com/ballle98/blink-sync-sentry"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
